requirements.txt
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
Deprecated==1.2.14
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycparser==2.21
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
tomli==2.0.1
urllib3==2.0.4
wrapt==1.15.0
